update overview API

+++ src/main/java/com/nemi/report/model/response/overview/BusinessTodayResponse.java
@@ -48,7 +48,7 @@ public class BusinessTodayResponse {
         private BigDecimal revenue;
 
         @JsonProperty("orders")
-        private BigDecimal orders;
+        private Long orders;
     }
 
     @Data
